refactor(examples): ReloadablePane moves to its own file (#17948)
